+++ src/main/java/com/dsm/nms/domain/comment/repository/CommentRepository.java
@@ -1,0 +1,7 @@
+package com.dsm.nms.domain.comment.repository;
+
+import com.dsm.nms.domain.comment.entity.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
+}
