+++ src/main/java/com/dsm/nms/domain/notice/repository/NoticeTargetRepository.java
@@ -1,0 +1,6 @@
+package com.dsm.nms.domain.notice.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NoticeTargetRepository extends JpaRepository<NoticeTarget, NoticeTargetId> {
+}
