+++ src/main/java/com/dsm/nms/domain/notice/repository/TargetRepository.java
@@ -1,0 +1,6 @@
+package com.dsm.nms.domain.notice.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TargetRepository extends JpaRepository<Target, Integer> {
+}
