+++ src/main/java/com/dsm/nms/domain/reply/repository/ReplyRepository.java
@@ -1,0 +1,7 @@
+package com.dsm.nms.domain.reply.repository;
+
+import com.dsm.nms.domain.reply.entity.Reply;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReplyRepository extends JpaRepository<Reply, Integer> {
+}
