--- src/main/java/com/dsm/nms/domain/star/StarRepository.java
@@ -1,6 +1,0 @@
-package com.dsm.nms.domain.star;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StarRepository extends JpaRepository<Star, Integer> {
-}
