+++ src/main/java/com/dsm/nms/domain/star/repository/StarRepository.java
@@ -1,0 +1,7 @@
+package com.dsm.nms.domain.star.repository;
+
+import com.dsm.nms.domain.star.entity.Star;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StarRepository extends JpaRepository<Star, Integer> {
+}
