--- src/main/java/com/dsm/nms/domain/student/Grade.java
@@ -1,5 +1,0 @@
-package com.dsm.nms.domain.student;
-
-public enum Grade {
-    FIRST, SECOND, THIRD
-}
