--- src/main/java/com/dsm/nms/domain/student/StudentRepository.java
@@ -1,9 +1,0 @@
-package com.dsm.nms.domain.student;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface StudentRepository extends JpaRepository<Student, Integer> {
-    Optional<Student> findByEmail(String email);
-}
