+++ src/main/java/com/dsm/nms/domain/student/entity/Grade.java
@@ -1,0 +1,5 @@
+package com.dsm.nms.domain.student.entity;
+
+public enum Grade {
+    FIRST, SECOND, THIRD
+}
