+++ src/main/java/com/dsm/nms/domain/student/repository/StudentRepository.java
@@ -1,0 +1,10 @@
+package com.dsm.nms.domain.student.repository;
+
+import com.dsm.nms.domain.student.entity.Student;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface StudentRepository extends JpaRepository<Student, Integer> {
+    Optional<Student> findByEmail(String email);
+}
