--- src/main/java/com/dsm/nms/domain/teacher/TeacherRepository.java
@@ -1,9 +1,0 @@
-package com.dsm.nms.domain.teacher;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
-    Optional<Teacher> findByUsername(String username);
-}
